Cast counts and sums to integers
Prevents sequelize from resolving as string

+++ server/migrations/20200717051735-create-materialized-view-bgatlas2008-observed-species.js
@@ -10,7 +10,7 @@ module.exports = {
       SELECT
         bgatlas2008_utm_code as utm_code,
         o.species,
-        SUM(count) as count,
+        SUM(count)::integer as count,
         (count(e.species) > 0) as existing
       FROM birds_observations o
       LEFT JOIN bgatlas2008_species e ON (e.utm_code = o.bgatlas2008_utm_code AND e.species = o.species)
