+++ server/migrations/20200717073453-create-view-bgatlas2008-stats-global.js
@@ -16,9 +16,9 @@ module.exports = {
       LEFT JOIN (
         SELECT
           utm_code,
-          sum(case when existing then 1 else 0 end) as spec_known,
-          sum(case when existing then 0 else 1 end) as spec_unknown,
-          (SELECT count(e.species) FROM bgatlas2008_species e WHERE o.utm_code = e.utm_code) as spec_old
+          sum(case when existing then 1 else 0 end)::integer as spec_known,
+          sum(case when existing then 0 else 1 end)::integer as spec_unknown,
+          (SELECT count(e.species)::integer FROM bgatlas2008_species e WHERE o.utm_code = e.utm_code) as spec_old
         FROM bgatlas2008_observed_species o
         GROUP BY o.utm_code
       ) s ON (c.utm_code = s.utm_code)
